# [Enhancement][FlatJson] opitmize flat json compaction performance