--- a/be/src/storage/rowset/json_column_compactor.h
+++ b/be/src/storage/rowset/json_column_compactor.h
@@ -32,6 +32,10 @@ public:
 
 private:
     Status _compact_columns(std::vector<ColumnPtr>& json_datas);
+
+    Status _merge_columns(std::vector<ColumnPtr>& json_datas);
+
+    Status _flatten_columns(std::vector<ColumnPtr>& json_datas);
 };
 
 class JsonColumnCompactor final : public ColumnWriter {