--- a/be/src/storage/rowset/json_column_writer.cpp
+++ b/be/src/storage/rowset/json_column_writer.cpp
@@ -92,6 +92,8 @@ Status FlatJsonColumnWriter::_flat_column(std::vector<ColumnPtr>& json_datas) {
     _flat_types = deriver.flat_types();
     _has_remain = deriver.has_remain_json();
 
+    VLOG(1) << "FlatJsonColumnWriter flat_column flat json: "
+            << JsonFlatPath::debug_flat_json(_flat_paths, _flat_types, _has_remain);
     if (_flat_paths.empty()) {
         return Status::InternalError("doesn't have flat column.");
     }
@@ -281,6 +283,7 @@ Status FlatJsonColumnWriter::finish_current_page() {
 StatusOr<std::unique_ptr<ColumnWriter>> create_json_column_writer(const ColumnWriterOptions& opts,
                                                                   TypeInfoPtr type_info, WritableFile* wfile,
                                                                   std::unique_ptr<ScalarColumnWriter> json_writer) {
+    VLOG(1) << "Create Json Column Writer is_compaction: " << opts.is_compaction << ", need_flat : " << opts.need_flat;
     // compaction
     if (opts.is_compaction) {
         if (opts.need_flat) {