… be/src/storage/rowset/json_column_iterator.cpp
… be/src/storage/tablet_reader.cpp
… be/src/util/json_flattener.cpp
… be/src/util/json_flattener.h
… be/test/storage/rowset/flat_json_column_compact_test.cpp
… be/test/storage/rowset/flat_json_column_rw_test.cpp