## Why I'm doing:

## What I'm doing:
Follow PR #48104

this pr for optimize compaction performance
| SSB | Scalar Data | Json | Flat Json(Before #48104) | FlatJson（PR #48104) | FlatJson (This PR) |
| -- | -- | -- | -- | -- | -- |
| Data Size | 58.828 GB | 152.616GB | 245.947 GB | 58.571 GB | 58.571 GB |
| compaction time | 4min30s | 35min20s | 46min51s | 23 min 30s | 6min12s |

Compaction Before:
Read all json subfield in FlatJson -> Merge to a real json -> Compaction merge rows base on json -> json flatten to flat-subfield -> storage flat json

Compaction Now:
To derive the flat json schema when read -> Read all json subfield by schema -> Compaction merge rows base on same flat json schema  -> to derive new schema base on flat json -> storage flat json

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5